requirements.txt
openai~=1.70.0
python-dotenv~=1.1.0
pyttsx3~=2.98
sounddevice~=0.5.1
scipy~=1.15.2
pynput~=1.8.1
numpy~=1.26.4
torch~=2.2.2
pydub~=0.25.1
torchaudio~=2.2.2